# perf(treeData): decrease Tree Grid expandAll/collapseAll time by 40x